requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
rembg==2.0.66
Pillow==10.0.1
numpy==1.24.3
requests==2.31.0
gunicorn==21.2.0 